+++ pyanaconda/modules/payload/payloads/live_image/installation.py
